requirements.txt
opencv-python-headless==4.8.1.78
numpy==1.26.2
picamera2==0.3.16
ultralytics==8.0.227
pyyaml==6.0.1
python-dotenv==1.0.0
pyaudio==0.2.13
SpeechRecognition==3.10.0
wave==0.0.2
